Split delete comments out into it's own decorator
