+++ lib/api.js
@@ -47,7 +47,8 @@ module.exports = settings => {
 
   flow.decorate(decorators.metadata(settings));
   flow.decorate(decorators.deadline(settings));
-  flow.decorate(decorators.comments(settings), { list: false });
+  flow.decorate(decorators.filterComments(settings), { list: false });
+  flow.decorate(decorators.deleteComments(settings), { list: false });
   flow.decorate(decorators.activityLog(settings), { list: false });
   flow.decorate(decorators.nextSteps(settings), { list: false });
   flow.decorate(decorators.licenceHolder(settings), { list: false });
