+++ lib/decorators/filter-comments.js
@@ -5,8 +5,6 @@ module.exports = settings => {
       return c;
     }
 
-    const deletedComments = c.activityLog.filter(log => log.eventName === 'delete-comment').map(log => log.event.meta.id);
-
     const statusChanges = c.activityLog
       .filter(a => a.eventName.match(/^status:/))
       .sort((a, b) => {
@@ -59,16 +57,6 @@ module.exports = settings => {
           return {
             ...activity,
             isNew: isNew(activity)
-          };
-        }
-        return activity;
-      })
-      .map(activity => {
-        if (activity.eventName === 'comment' && deletedComments.includes(activity.id)) {
-          return {
-            ...activity,
-            deleted: true,
-            comment: null
           };
         }
         return activity;
