+++ lib/decorators/index.js
@@ -4,5 +4,6 @@ module.exports = {
   nextSteps: require('./next-steps'),
   licenceHolder: require('./licence-holder'),
   deadline: require('./deadline'),
-  comments: require('./filter-comments')
+  filterComments: require('./filter-comments'),
+  deleteComments: require('./delete-comments')
 };
